+++ src/lib.rs
@@ -7,6 +7,7 @@ pub mod clock;
 pub mod gpio;
 pub mod i2c;
 pub mod prelude;
+pub mod pwm;
 pub mod spi;
 pub mod time;
 pub mod timer;